Rename file as well
