+++ packages/ui/src/app/pages/Proposals/PastProposals.tsx
@@ -6,7 +6,7 @@ import { FilterPageHeader } from '@/common/components/forms/FilterBox'
 import { MainPanel } from '@/common/components/page/PageContent'
 import { SearchProcess } from '@/common/components/page/SearchProcess'
 import { SidePanel } from '@/common/components/page/SidePanel'
-import { useMockActivities } from '@/common/hooks/useActivities'
+import { useMockActivities } from '@/common/hooks/useMockActivities'
 import { AddProposalButton } from '@/proposals/components/AddProposalButton'
 import { ProposalEmptyFilter, ProposalFilters } from '@/proposals/components/ProposalFilters'
 import { ProposalList } from '@/proposals/components/ProposalList'
