+++ packages/ui/src/app/pages/Proposals/Proposals.tsx
@@ -7,7 +7,7 @@ import { MainPanel } from '@/common/components/page/PageContent'
 import { PageTitle } from '@/common/components/page/PageTitle'
 import { SearchProcess } from '@/common/components/page/SearchProcess'
 import { SidePanel } from '@/common/components/page/SidePanel'
-import { useMockActivities } from '@/common/hooks/useActivities'
+import { useMockActivities } from '@/common/hooks/useMockActivities'
 import { AddProposalButton } from '@/proposals/components/AddProposalButton'
 import { NoProposals } from '@/proposals/components/NoProposals'
 import { ProposalList } from '@/proposals/components/ProposalList'
